requirements.txt
tkinter
time
datetime
send2trash
cryptography
requests
beautifulsoup4
os
shutil
random
pillow